+++ internal/service/issue_relation_info.go
@@ -6,6 +6,7 @@ import (
 	"tirelease/commons/git"
 	"tirelease/internal/dto"
 	"tirelease/internal/entity"
+	"tirelease/internal/repository"
 )
 
 // ================================================================ Compose Function From Remote Query
@@ -38,14 +39,24 @@ func ComposeTriageRelationInfoByIssue(issue *entity.Issue) (*dto.IssueRelationIn
 }
 
 func ComposeIssueAffectsByIssue(issue *entity.Issue) (*[]entity.IssueAffect, error) {
+	// search all minor version
+	releaseVersionOption := &entity.ReleaseVersionOption{
+		Type:   entity.ReleaseVersionTypeMinor,
+		Status: entity.ReleaseVersionStatusOpen,
+	}
+	releaseVersions, err := repository.SelectReleaseVersion(releaseVersionOption)
+	if nil != err {
+		return nil, err
+	}
+
+	// init all affects of issue
 	issueAffects := make([]entity.IssueAffect, 0)
-	// todo: minorVersion is fixed now, need to search later - 2022.2.7
-	for _, minorVersion := range MinorVersionList {
+	for _, releaseVersion := range *releaseVersions {
 		issueAffect := &entity.IssueAffect{
 			CreateTime:    time.Now(),
 			UpdateTime:    time.Now(),
 			IssueID:       issue.IssueID,
-			AffectVersion: minorVersion,
+			AffectVersion: releaseVersion.Name,
 			AffectResult:  entity.AffectResultResultUnKnown,
 		}
 		issueAffects = append(issueAffects, *issueAffect)
@@ -97,4 +108,121 @@ func ComposeIssuePrRelationsByIssue(issue *entity.Issue) (*[]entity.IssuePrRelat
 
 	// Return
 	return &issuePrRelations, &pullRequests, nil
+}
+
+// ============================================================================ CURD Of IssueRelationInfo
+func SaveIssueRelationInfo(triageRelationInfo *dto.IssueRelationInfo) error {
+	// Save Issue
+	if err := repository.CreateOrUpdateIssue(triageRelationInfo.Issue); nil != err {
+		return err
+	}
+
+	// Save IssueAffects
+	for _, issueAffect := range *triageRelationInfo.IssueAffects {
+		if err := repository.CreateIssueAffect(&issueAffect); nil != err {
+			return err
+		}
+	}
+
+	// Save IssuePrRelations
+	for _, issuePrRelation := range *triageRelationInfo.IssuePrRelations {
+		if err := repository.CreateIssuePrRelation(&issuePrRelation); nil != err {
+			return err
+		}
+	}
+
+	// Save PullRequests
+	for _, pullRequest := range *triageRelationInfo.PullRequests {
+		if err := repository.CreateOrUpdatePullRequest(&pullRequest); nil != err {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func SelectIssueRelationInfo(option *dto.IssueRelationInfoQuery) (*[]dto.IssueRelationInfo, error) {
+	// Select Issues
+	issueOption := &entity.IssueOption{
+		ID:      option.ID,
+		IssueID: option.IssueID,
+		Number:  option.Number,
+		State:   option.State,
+		Owner:   option.Owner,
+		Repo:    option.Repo,
+	}
+	issues, err := repository.SelectIssue(issueOption)
+	if nil != err {
+		return nil, err
+	}
+
+	// From Issue to select IssueAffects & IssuePrRelations & PullRequests
+	alls := make([]dto.IssueRelationInfo, 0)
+	for _, issue := range *issues {
+		issueRelationInfo, err := ComposeRelationInfoByIssue(&issue)
+		if nil != err {
+			return nil, err
+		}
+		alls = append(alls, *issueRelationInfo)
+	}
+
+	// Filter & Result
+	issueRelationInfos := make([]dto.IssueRelationInfo, 0)
+	for _, issueRelationInfo := range alls {
+		var filter bool = false
+		for _, issueAffect := range *issueRelationInfo.IssueAffects {
+			if option.AffectVersion == "" || issueAffect.AffectVersion == option.AffectVersion {
+				filter = true
+				break
+			}
+		}
+		if filter {
+			issueRelationInfos = append(issueRelationInfos, issueRelationInfo)
+		}
+	}
+
+	return &issueRelationInfos, nil
+}
+
+// ============================================================================ Inner Function
+func ComposeRelationInfoByIssue(issue *entity.Issue) (*dto.IssueRelationInfo, error) {
+	// Find IssueAffects
+	issueAffectOption := &entity.IssueAffectOption{
+		IssueID: issue.IssueID,
+	}
+	issueAffects, err := repository.SelectIssueAffect(issueAffectOption)
+	if nil != err {
+		return nil, err
+	}
+
+	// Find IssuePrRelations
+	issuePrRelationOption := &entity.IssuePrRelationOption{
+		IssueID: issue.IssueID,
+	}
+	issuePrRelations, err := repository.SelectIssuePrRelation(issuePrRelationOption)
+	if nil != err {
+		return nil, err
+	}
+
+	// Find PullRequests
+	pullRequests := make([]entity.PullRequest, 0)
+	for _, issuePrRelation := range *issuePrRelations {
+		pullRequestOption := &entity.PullRequestOption{
+			PullRequestID: issuePrRelation.PullRequestID,
+		}
+		pullRequest, err := repository.SelectPullRequestUnique(pullRequestOption)
+		if nil != err {
+			return nil, err
+		}
+		pullRequests = append(pullRequests, *pullRequest)
+	}
+
+	// Construct IssueRelationInfo
+	issueRelationInfo := &dto.IssueRelationInfo{
+		Issue:            issue,
+		IssueAffects:     issueAffects,
+		IssuePrRelations: issuePrRelations,
+		PullRequests:     &pullRequests,
+	}
+	return issueRelationInfo, nil
 }
