+++ internal/service/version_triage_info_test.go
@@ -1,0 +1,31 @@
+package service
+
+import (
+	"testing"
+	"time"
+
+	"tirelease/commons/database"
+	"tirelease/commons/git"
+	"tirelease/internal/entity"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestCreateOrUpdateVersionTriageInfo(t *testing.T) {
+	git.Connect(git.TestToken)
+	git.ConnectV4(git.TestToken)
+	database.Connect(generateConfig())
+
+	versionTriage := &entity.VersionTriage{
+		CreateTime:   time.Now(),
+		UpdateTime:   time.Now(),
+		VersionName:  "5.4.0",
+		IssueID:      git.TestIssueNodeID,
+		TriageResult: entity.VersionTriageResultAccept,
+	}
+	info, err := CreateOrUpdateVersionTriageInfo(versionTriage)
+
+	assert.Equal(t, true, err == nil)
+	assert.Equal(t, true, info != nil)
+	assert.Equal(t, true, info.IsAccept)
+}
